Production code: Added watchdog. Changed timeouts to rely on millis() based variables.
This code is running on wind machine exhibit as of 2/18/2013. I'm not
entirely sure where the '-999' was creeping in to the system but it's
been removed. All three RPMs are not reset after 30 seconds and 10
minutes. Also added watchdog timer to make sure system never gets stuck.
resetPeak was modified heavily - now values reduce to zero at the same
rate no matter what level they are at.